+++ Changelog.md
@@ -12,6 +12,8 @@
 - Updated rv to v0.10
 - Changed the way prediction is done on continuous columns to achieve a 4x
     speedup.
+- Changed the way continuous-categorical mutual information is computed for a
+    40x speedup
 
 ## 0.32.3
 - Add '--flat-columns' argument to `braid run` command so that an engine can be